+++ packages/config/maintenance/lintConfigFiles.ts
@@ -20,7 +20,11 @@ import {
 } from "../src/devices/DeviceConfig";
 import type { DeviceID } from "../src/devices/shared";
 import { parseLogic } from "../src/Logic";
-import { configDir, getDeviceEntryPredicate } from "../src/utils";
+import {
+	configDir,
+	getDeviceEntryPredicate,
+	versionInRange,
+} from "../src/utils";
 
 const configManager = new ConfigManager();
 
@@ -1079,13 +1083,27 @@ Consider converting this parameter to unsigned using ${white(
 		if (typeof other === "boolean" || typeof me === "boolean") {
 			if (other !== me) continue;
 		} else {
-			if (other.min !== me.min || other.max !== me.max) continue;
+			// Ensure that the firmware version ranges do not overlap,
+			// except if one is preferred and the other isn't
+			if (
+				versionInRange(me.min, other.min, other.max) ||
+				versionInRange(me.max, other.min, other.max) ||
+				versionInRange(other.min, me.min, me.max) ||
+				versionInRange(other.max, me.min, me.max)
+			) {
+				if (entry.preferred !== index[firstIndex].preferred) {
+					continue;
+				}
+			} else {
+				continue;
+			}
 		}
 		// This is a duplicate!
 		addError(
 			entry.filename,
-			`Duplicate config file detected for device (manufacturer id = ${entry.manufacturerId}, product type = ${entry.productType}, product id = ${entry.productId})
-The first occurrence of this device is in file config/devices/${index[firstIndex].filename}`,
+			`Duplicate config file detected for device (manufacturer id = ${entry.manufacturerId}, product type = ${entry.productType}, product id = ${entry.productId}, firmware range ${me.min} to ${me.max})
+The first occurrence of this device is in file config/devices/${index[firstIndex].filename}, firmware range ${other.min} to ${other.max}.
+If this is intended, consider marking one of the config files as preferred or split files by firmware version.`,
 		);
 	}